CRAS: audio_thread - Modify init_cb_ts to avoid hw_level rise
If we set the init_cb_ts of the new input stream to now +
sleep_interval_ts, the hw_level of input device will rise. It is because
the stream need to wait until next_cb_ts to read no matter it already
has enough samples or not.

We can set init_cb_ts to the time when stream gets enough samples to
post. It can resolve this issue.

BUG=chromium:912457
TEST=Can pass audio_CrasOutputStress.input_only test on peppy relibly.

Change-Id: I4bf6598438647cdbbe64c91c89999a033d05e109
Reviewed-on: https://chromium-review.googlesource.com/1369344
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>