CRAS: audio_thread - Append new streams after consuming remaining frames
Because there are still the remaining samples in the device after
removing streams. If we add new output stream immediately, the new
samples will append to old samples. It will cause the hardware level
rise.

To fix it, we can set next_cb_ts to the time when hw_level of device
is close to min_cb_level of the new stream.

BUG=chromium:912457
TEST=Can pass audio_CrasOutputStress.output_only test on peppy relibly.

Change-Id: Id947af6b241f3266f85ee3b8f7541c925328ec3d
Reviewed-on: https://chromium-review.googlesource.com/1517214
Commit-Ready: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Tested-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>