CRAS: Add a new field to represent when pending streams will start
If one stream is attached to two devices, both iodevs will use the same
next_cb_ts. Suppose one is a pending stream and another is a running
stream. If audio thread fetchs the running stream first, it will cause
next_cb_ts to be changed so that the pending stream will not be moved
into running streams list forever.

To fix it, we need to add new field init_cb_ts, which is the time when
pending stream should be moved to running stream. And then we don't need
to worry that next_cb_ts is changed from other dev_stream which has the
same rstream.

BUG=b:121339677
TEST=Audio works fine after changing playback devices.

Change-Id: I0bb76344cece38e1076e05ed882303f87ccbfd4a
Reviewed-on: https://chromium-review.googlesource.com/1391928
Commit-Ready: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Tested-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>