CRAS: HFP - Remove code to dynamically update packet and buffer size
Change the main buffer size to 28800 so it's multiple of packet
size values we've seen(48, 60, and 128).

The old mechanism assumes the SCO packet come in arbitrary size
so we have to dynamically change the iodev's buffer_size since it
might not divide the main large buffer. The dangerous thing is that
the packet size is determined in audio thread, and it may trigger
the buffer size change on an iodev when main thread still using it.

Thre are only handful of packet size values, it's better we set the
main buffer size as their common multiple.

BUG=chromium:946356
TEST=Manual test HFP audio

Change-Id: Ib7b3fa3411f3526fcfa849eabcc2d2c705491857
Reviewed-on: https://chromium-review.googlesource.com/1517343
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Hsinyu Chao <hychao@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>