CRAS: Add new flag to indicate if dev_stream starts fetching
When we add the new stream, the next_cb_ts of this stream would be set
to the earliest next_cb_ts of others. It causes the new stream to remain
empty before next fetch time. It also means we can not write any frames
during this time. However, if we use the big client buffer and the
small device buffer, the next fetch time is far from now. It will
trigger underrun.

To fix this issue, we can add the new flag in dev_stream to indicate
if dev_stream starts fetching. The audio thread will skip mixing streams
which have not started running yet. So that the new stream will not affect
other existing streams.

BUG=chromium:916887
TEST=Playing two streams does not trigger underruns on grunt.

Change-Id: Ida6d6eb92784b9f3abebcd8456eda346dc89c8a7
Reviewed-on: https://chromium-review.googlesource.com/1510895
Commit-Ready: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Tested-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>