CRAS: Add JackType to specify a node is always plugged
The original code only set plugged on the first device of the USB card.
However, this logic has its limitation as there is a dock having both
headphone and lineout, but no jack and no UAC2/UAC3 connector status
support.

To fix the problem, we do not consider the is_first condition for
fully-specified UCM and add the new JackType "always" to specify that
the node is always plugged even without jack.

BUG=chromium:1524611
TEST=Can set JackType "always" on HiFi.conf of USB device.
CQ-DEPEND=CL:1530352

Change-Id: Ie5a63c75558cc6556ecc7deda293159352da1943
Reviewed-on: https://chromium-review.googlesource.com/1530351
Commit-Ready: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Tested-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>