CRAS: loopback_iodev - Fix bug when output device changes
Currently loopback iodev re-registers to the active output at
device enabled event, and there's a bug of that when user switches
output from A to B. But CRAS enables the fallback output device
for this output transition so loopback will mistakenly be
registered to the fallback output.

Fix this problem by adding loopback register to device disable
event as well.

BUG=chromium:946356
TEST=cras_test_client --loopback /tmp/1
Play youtube and switch output from A to B, verify that
all audio gets recorded in /tmp/1

Change-Id: I7fc61791e678bcd009b8f726b1b9e087e214be61
Reviewed-on: https://chromium-review.googlesource.com/1532388
Commit-Ready: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>