CRAS: cras_iodev - Open buf_state when adding stream
In CL:1510895, we changed the logic to create buf_state after
stream's first fetch. But we forgot to check whether buf_state
exists when we remove streams. It might make CRAS crash.

The better solution is creating buf_state in the first add_stream
because we destroy buf_state in the last remove_stream. Hence
buf_state always exists if there is a stream in iodev.

BUG=chromium:916887
TEST=Run CrasStress with bluetooth doesn't make CRAS crash.

Change-Id: Idbaf3c29ee99004a5f82a58b578d9c980d011b0a
Reviewed-on: https://chromium-review.googlesource.com/1535299
Commit-Ready: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Tested-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>