CRAS: iodev - Remove ext_format
The ext_format of iodev was introduced 4 years ago by commit:
commit 2db05743b CRAS: iodev - Allow DSP to modify number of channels

whose purpose was to allow DSP like beamforming to change channel
count. But we don't have a real use case so far, which means these two
variables are just always identical. Plus that the code has grown
quite different since then, keeping two member variables 'format' and
'ext_format' becomes confusing and hard to maintain. It's not trivial
to tell when to use which 'format'. Let's remove it for now, and
re-think how this should be implemented if in future we have a real
DSP use case that changes channel count.

BUG=chromium:946356
TEST=unittest

Change-Id: I6966d70086e4b068440d757bd24041a4f87ebc4e
Reviewed-on: https://chromium-review.googlesource.com/1573068
Commit-Ready: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>