CRAS: Update the .clang-format ForEachMacros
We copy the .clang-format directly from Linux kernel, which contains the
ForEachMacros for Linux keral instead of for CRAS. Update the field and
the command to get it. We might need to update the command or the names
of our ForEachMacros to better cover them if we find any missed one.

BUG=chromium:950858
TEST=None

Change-Id: I161ac3013001ca6eb4088bd9579bf8f36717a4e4
Reviewed-on: https://chromium-review.googlesource.com/1577391
Commit-Ready: En-Shuo Hsu <enshuo@google.com>
Tested-by: En-Shuo Hsu <enshuo@google.com>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>