CRAS: cras_server_metrics - Check thread before sending messages
Before sending CRAS main messages, we should check which thread we
belong to. If we are in the main thread, we can handle it directly.
It is a precursor to enable logging UMA to the main thread.

If the size of message is greater than PIPE_BUF, which is 4096 bytes in
linux, it will be split into multiple packets. It may cause race
condition if we send messages from main thread and audio thread an the
same time. The CL can also avoid this situation.

BUG=chromium:960297
TEST=UMA logs can be recorded well on peppy.

Change-Id: Iad650e73da10489cafa0c64a59bbe5ae2660a7a6
Reviewed-on: https://chromium-review.googlesource.com/1598997
Commit-Ready: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Tested-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>