CRAS: iodev_list - Do not ramp volume when muted
If volume is changed while muted, volume ramps are still applied.
If audio is playing when these ramps are applied, it is briefly audible
even though the device is supposed to be muted.

To fix this, check that the system is not muted before adding a
volume ramp.

BUG=b:131579153
TEST=set volume to 50%, start audio playback, hit mute key, hit volume
     down key. No sound should be heard.

Change-Id: I49ce98b571ec23cb16cb77be0323b6f3f8764a93
Reviewed-on: https://chromium-review.googlesource.com/1625624
Commit-Ready: Fletcher Woodruff <fletcherw@chromium.org>
Tested-by: Fletcher Woodruff <fletcherw@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>