cras: Update Cras API to work with btdispatch
Currently Cras API are working with bluez instead of btdispatch.
To correct this, we need to (1) add forwarding rules to bluez,
(2) change the permission of cras. (3) change the service name
defined in cras
This patch only contains the code changed in cras side, which
is (3).

BUG=None
TEST=test it as follows
1. start btdispatch
2. pair & connect BOSE QC35 with DUT
3. verify the headset works fine(play sound, increase/decrease volumn)
4. stop btdispatch
5. try to connect BOSE QC35 with DUT, but failed.
6. do 1-3 again
Cq-Depend: chromium:1673958
Change-Id: I9f2ab27862ff1eada5af544300358c5c4e1dcd0e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/1673767
Tested-by: Yun-Hao Chung <howardchung@google.com>
Commit-Queue: Yun-Hao Chung <howardchung@google.com>
Reviewed-by: Miao-chen Chou <mcchou@chromium.org>