CRAS: dev_io - Add the overrun op threshold
On grunt, the number of frames an input node reads once is decided by a
period size, whose default vaule is 1024.
According to the latest overrun ops, if the device's hw_level is larger
than 2 * largest_cb_level, the audio thread will drop frames.
However, the cb_level of capture streams are 256 from ARC++ and 480 from
Chrome. Both of them * 2 are less than 1024, it causes the overrun ops
will be triggered every time.

To fix the problem, we can add the overrun op threshold. The op only
works when the hw_level is larger than that threshold. That time must
to be longer than 2048 frames with rate 48000 (grunt only supports
48000 rate). Therefore, we think 50ms is a reasonable threshold.

For users, it means that the delay will be reduced only if it is larger
than 50ms. The value is too small to affect the user experience.

Notice that we can not fix the problem by changing a period size on
grunt because it will change a buffer size, too.
(Buffer size = 2 * period size on grunt. See #22 on b/112860209)

BUG=b:138076270, chromium:918462
TEST=The capture frames is not dropped on grunt and the delay keeps in 50ms
on panther.

Change-Id: Ibe367a2075fb6a49b45cca1449c895321c824585
Reviewed-on: https://chromium-review.googlesource.com/1708817
Tested-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Commit-Ready: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>