CRAS: hfp_slc - Fix AT+BCS reply OK
According to HFP 1.7.1 spec
The AG shall always respond with OK if the ID received in
AT+BCS is the same as the one sent, and with ERROR otherwise.

Airpod headset is affected by the bug. Fix it so that OK is
sent as reply to AT+BCS= command before we initialize HFP AG.

BUG=chromium:980454
TEST=Manual test Airpod in wideband speech mode.

Change-Id: I6036385c896b10796e6ba1c78d28fb5694dab2ee
Reviewed-on: https://chromium-review.googlesource.com/1716472
Tested-by: Hsinyu Chao <hychao@chromium.org>
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: En-Shuo Hsu <enshuo@google.com>