CRAS: dev_io - Reduce the threshold of dropping frames
Originally, the overrun ops drop frames when hw_level is larger than 2 *
largest_cb_level. However, the buffer size is only 8192 on Kabylake.
If the largest_cb_level is 4800 (assistant), the overrun ops will never
be triggered. Therefore, we change the ratio to 1.5 in this change.

BUG=chromium:918462
TEST=The capture works fine on eve.

Change-Id: Ie46cf7edf873ec28a0af64fea01bb72ffb916927
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/1782238
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Commit-Queue: Yu-Hsuan Hsu <yuhsuan@chromium.org>