CRAS: a2dp - Remove pre-fill zero frames
We pre-filled zero frames before the first data put to
A2DP iodev, the purpose is to stabilize the buffer level
at A2DP use case. This code was added a long time ago
and we think it's now okay to be removed.

Few facts to justify the pre-filling is not needed:
 - Audio thread scheduling has grown a lot more stable
then before.
 - no_stream ops was added to A2DP iodev, so some zero
frames are used at the transition of two streams.

Most importantly, we have discovered that the pre-filling
of zero frames at the beginning leads to L/R sync issue on
specific BT headset. Probably because the burst of samples
breaks the calibration in L and R pieces of headset.

If this change causes underrun regression, we should use
iodev's output_underrun callback to address that.

BUG=chromium:967038
TEST=Listen to YouTube by Airpod

Change-Id: I96286a11a028915ea00ddf27a9a89894e91eaf84
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/1788970
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>