CRAS: transport - Destroy after endpoint is suspended
Transport and endpoint are the objects defined in BlueZ media
API that represents A2DP headset connection. CRAS code was build
with the assumption that when a headset is disconnected, below
events are received in order:

(1) ClearConfiguration(transport T)
(2) Remove interface (transport T)

And we're now observing cases of reversed order, which causes
CRAS crash. Fix that by adding additional flag to mark if a
transport object has been removed by BT, and delay the destroy
call to after both events are received.

BUG=chromium:1001906
TEST=Pair two A2DP headsets A and B. From BT settings UI, connect
A and then proceed to connect B. The later connection would fail,
and now disconnect A. Repeat this with A/B swapped, verify that
CRAS does not crash.

Change-Id: I4d2f500bba774b00f4f0bf07bdac5d376b944550
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/1792422
Reviewed-by: En-Shuo Hsu <enshuo@google.com>
Tested-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>