CRAS: cras_iodev_list - Fix using multi-nodes on a single iodev
There are some problems about using multi-nodes on a single iodev.

(1)
If a stream pins to node 3:0 while UI calls select node 3:1 there
will be a problem in iodev list. In this scenario we try to do the
standard disable -> update_node -> enable sequence, but the pinned
stream causes iodev#3 still in use by audio thread, and that's a
potential multi thread problem. Besides, changing UCM control during
device running is not guaranteed to work.

(2)
If an iodev is diabled but there was pinned stream on it, no
iodev will serve this stream until it's pinned dev gets enabled
again. Fix this by allowing it pins to fallback iodev.

(3)
When node unplugged with a pinned stream attached, we should use
force disable call. Otherwise this device will be placed in a
bad state - not in enabled dev list, but used by audio thread.

BUG=b:141224765
TEST=Changing nodes doesn't break the pin stream on cave.

Change-Id: I125d574c755901d0597189ab077e611b93646ade
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/1816018
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>