CRAS: format cras_alsa_ucm.c
The clang-format repohook uses is 8 but the one in cros_sdk is 9 and
this difference seems to cause this minor difference. Format the file
with the cros_sdk clang-format to prevent it from breaking the build
after we enable the format check under src_test. This difference should be
rare as the repohook will only check the changed lines.
People should follow the src_test and just ignore the repohook while updating
the change if it is the case.

BUG=chromium:950858
TEST=FEATURES="test" USE=asan emerge-{BOARD} -v media-sound/adhd

Change-Id: Ibeddc1a3f787c376c13f224bfdb7f0c4b311373e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/1824620
Tested-by: En-Shuo Hsu <enshuo@google.com>
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Commit-Queue: En-Shuo Hsu <enshuo@google.com>