CRAS: cras_tests: add WAVE playback and capture
Create structs WavSource and WavSink which adapt hound::WavReader and
hound::WavWriter to implement the Read and Write traits, and use these
to implement WAVE playback and capture.

Dynamically select which source type to use based on the input/output
file name.

The audio is limited to integer sample formats, but supports all bit
depths that CRAS supports. S24_LE samples are upsampled to S32_LE due
to a quirk of the encoder used.

BUG=chromium:1008990
TEST="sox -n -r 48000 -b 16 output.wav synth 5 sine 330 &&
     cras_tests playback output.wav", check that sine wave is played
     accurately.
TEST="cras_tests capture output.wav && aplay output.wav",
     check that data is recorded properly.

Change-Id: I88f8080651305cad89c179cb00b5c327c625aff6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/1832584
Tested-by: Fletcher Woodruff <fletcherw@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Commit-Queue: Fletcher Woodruff <fletcherw@chromium.org>