CRAS: libcras: Add CrasServerState to CrasClient
Fix the code for CrasServerState so it can be properly mapped, and
add an instance of it to CrasClient, which will be initialized using
the server state fd returned in the CRAS_CLIENT_CONNECTED message.

Add two VolatileRefs to CrasServerState which can be used to read the
current system volume and mute level.

BUG=chromium:1008990
TEST=cargo test

Change-Id: Ie4483603634a428df607cd669baade67962df193
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/1836474
Reviewed-by: Fletcher Woodruff <fletcherw@chromium.org>
Tested-by: Fletcher Woodruff <fletcherw@chromium.org>
Commit-Queue: Fletcher Woodruff <fletcherw@chromium.org>