CRAS: cras_tests: add sigint handler
Add a handler for SIGINT which flips an atomic flag variable, and stop
playback/capture if that flag is flipped. If the flag has already been
flipped, the program is terminated via exit() to prevent the program
from hanging in a hard-to-kill state.

This way, we don't need explicit flush statements during capture since
buffers can be flushed gracefully when writers are dropped during exit.

BUG=chromium:1008990
TEST="cras_tests capture in.raw" and check that the recorded file
     contains all data.

Cq-Depend: chromium:1836185
Change-Id: I3719fba5c7a3f4591dd8ecb419baa480d0d553d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/1835599
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Tested-by: Fletcher Woodruff <fletcherw@chromium.org>
Commit-Queue: Fletcher Woodruff <fletcherw@chromium.org>