CRAS: alsa_jack - Fix jack matching rule in UCM
Below is our existing logic of gpio jack matching:
(1) Create jack for each input evdev.
(2) Look up a UCM device by matching the 'JackName' label
with input evdev name.
(3) Once the UCM device is found, look for the value of
'PlaybackPCM' or 'CapturePCM' label according to direction.
(4) Compare the PCM value with 'hw:<card-name>,<jack-list-index>'
which was added in commit bc573a333

The last step is NOT always true, there are cases that PCM
name doesn't really contain the exact string as card name. And
that blocks us from passing these PCM values to snd_pcm_open().

In fact, part of step (4) is redundant because in (2) we found
the UCM device that matches jack name so we already know it's
the correct card to associate.

We could simplify step (4) to just match the device index in
PCM name. So that we can fix 'PlaybackPCM' and 'CapturePCM'
values in UCM files to the ones that compatible with ALSA
libraries.

BUG=b:141447866, chromium:1014208
TEST=Verify node enumeration still works on octopus board.

Change-Id: I3df0115a10def14ead6a93f49deadda218b1aad1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/1861477
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>