CRAS: alsa_ucm - Add DependentPCM flag in alsa UCM
This is a reland of commit ea9baa66

On some boards the Playback/Capture PCM names written in
UCM file is not correct that cannot be passed to snd_pcm_open()
And that's why we have reverted this change.

We're going to reland this change with a later commit to handle
the incorrect PCM names in some UCM files. Seperate change will
be landed to fix the root cause and fix those UCM files.

BUG=b:141447866
TEST=Test on Nami board, add capture stream pinned to headset mic
and at the same time record from internal mic. Verify that both
recording stream receives valid audio.

Cq-Depend: chromium:1862363
Change-Id: I8e32ed34de358a1da24c772c02f6101babe2b4c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/1862362
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>