CRAS: Support querying intrinsic volume
This CL introduces a new config variable called IntrinsicVolume under UCM
which will be used as a volume baseline for CRAS. Our goal is to have CRAS
adjust gains based on it instead of tuning gains specifically for different
applications and use cases. That is to say, we can know how much gain we need
to add if we know the intrinsic volume of a device and the target volume.

The benefits adding it are
- Saving partners’ and our time tuning the capture gain. What partner
needs to do is to run a standard measurement for the input device's
intrinsic volume and set the value in the UCM. CRAS should handle the
following works.
- Providing more consistent audio experience across all chromebook
as the captured gain are adjusted to meet a standard target even if the
underlying hardware is different.
- Supporting different gain per use case, which is easier to maintain and
doesn't require per board tuning.

BUG=chromium:1020969
TEST=FEATURES=test emerge-"${BOARD}" media-sound/adhd

Change-Id: I483b6f76a0ecbdfbae462564582297a1f1c88e7b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/1926068
Tested-by: En-Shuo Hsu <enshuo@chromium.org>
Commit-Queue: En-Shuo Hsu <enshuo@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>