CRAS: libcras: Add Dummy and Null ShmStream impls
Adds two implementations of the ShmStream interface.

The Mock implementation only reports a ServerRequest when triggered by
calling the trigger_callback_with_timeout() method. This is useful for
testing, so that no race conditions/flaky tests are created.

The Null implementation returns ServerRequests at regular intervals
based on the buffer size and frame rate, but ignores all audio data.
This is useful for providing to a client when we do not want them to
actually be able to playback or capture audio, but we still want
callbacks at regular intervals.

BUG=chromium:968724
TEST=cargo test

Cq-Depend: chromium:1790200
Change-Id: I414e249fee2e4eff679462feb7ed7b37dd831b07
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/1871758
Tested-by: Fletcher Woodruff <fletcherw@chromium.org>
Reviewed-by: Chih-Yang Hsia <paulhsia@chromium.org>
Commit-Queue: Fletcher Woodruff <fletcherw@chromium.org>