CRAS: build - Allow building rust lib locally
The server is now partially built with rust. Currently the build depends
on portage installing the rust section as a system library. That makes
building with make on a workstation difficult.

To easy out-of-chroot builds, allow and default to building the rust
portion locally.

Cq-Depend: chromium:1966282
Change-Id: I667bd94f9a4ca93c7b16b2d718f0e2f43d0c8cbc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/1968111
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Reviewed-by: Chih-Yang Hsia <paulhsia@chromium.org>
Tested-by: Dylan Reid <dgreid@chromium.org>
Commit-Queue: Chih-Yang Hsia <paulhsia@chromium.org>