CRAS: Deprecate system gain related API
The CL deprecates the API used to change system gain on CRAS and cleans
up codes used for supporting it. The main purpose for this change is to
simplify the logic related to capture gain so it becomes easier for
partner to tune and for us to maintain.

BUG=b:147080207
TEST=Build,deploy and test on DUT

Cq-Depend: chromium:2007029, chrome-internal:2531574, chrome-internal:2534153, chromium:2049923
Change-Id: Id03b296c4c9d52f584d36214277965316958ca79
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/2006931
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>
Tested-by: En-Shuo Hsu <enshuo@chromium.org>
Commit-Queue: En-Shuo Hsu <enshuo@chromium.org>