CRAS: Handle HOTWORD_STREAM on not hotword device
The change is for handling adding stream with HOTWORD_STREAM flag on a not
hotword device, which don't exist now but will do after crrev.com/c/1997003 is
landed. Ideally, we can consider to add a higher level abstract of usage
or feature flag instead of stream flag to reflect the use of hotwording,
but this is a simpler change that can have things work.

BUG=chromium:1020969
TEST=Build, deploy and check if hotwording still works without
crrev.com/c/1997003. And validate again with crrev.com/c/1997003.

Change-Id: Icfe99f1ab298c349efa0bd7c52174a48ee54b74e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/1994813
Reviewed-by: Chih-Yang Hsia <paulhsia@chromium.org>
Tested-by: En-Shuo Hsu <enshuo@chromium.org>
Commit-Queue: En-Shuo Hsu <enshuo@chromium.org>