CRAS: Make some struct helpers return instead of fill the struct
Instead of filling a pointer to the struct, create the struct and return
it. This prevents the compiler from issuing uninitialized code warnings
if the function doesn't get inlined.

Remove the usage of audio_format from cras_client as the mfmt variable
wasn't used.

Change-Id: I9182da50632b359ecffb0c2147a379965d389649
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/1999283
Tested-by: Dylan Reid <dgreid@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Commit-Queue: Dylan Reid <dgreid@chromium.org>