CRAS: dsp - leave existing ini if new ini fails
When reloading the DSP ini file, if creation of the new ini fails,
leave the existing ini in place.

This fixes at least the leak of the old ini struct, which wasn't freed
in the early return path.

In addition it _could_ help the intermittent fuzzer use-after-free
errors. Theoretically the fuzzer could call this multiple times and get
failures on some (maybe for out of FDs or other errors opening the
file). In that case the global ini is set to NULL. However, importantly,
the pipelines aren't destroyed. This leaves pointers to the ini in the
running pipelines, since they no longer match the global ini each
pipeline could try to free the same ini, causing use-after free.

BUG=1038527

Change-Id: Ie29572efa345b22846ff3684b5cad215dab8b91d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/2004291
Commit-Queue: Chih-Yang Hsia <paulhsia@chromium.org>
Tested-by: Chih-Yang Hsia <paulhsia@chromium.org>
Reviewed-by: Chih-Yang Hsia <paulhsia@chromium.org>