cras: remove JackType "always" support
In order to better align with upstream UCM we need to drop JackType
entirely from support and move to JackControl/JackDev. The "always" type
is a work around for USB devices that do not export a jack device. The
solution is to assume the jack is connected if the USB device is
connected if not jack is found. If users don't want that jack used then
they can change the device in the UI.

BUG=chromium:1041791
TEST=Test audio device selection when inserting a known device dependent
on "always" type

Cq-Depend: chromium:2023329
Change-Id: Ibff5e7282924f1055f686af54b6b998c703c84e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/2004264
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Tested-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Commit-Queue: Yu-Hsuan Hsu <yuhsuan@chromium.org>