CRAS: audio_format - Verify the stream layout for new streams
The stream layout can be invalid and have indices that are off the end
of the array. The fuzzer is hitting such a case. Check them before use.
In the future more checks can be moved here. For now do the minimum to
ensure memory safety.

BUG=oss-fuzz:20163
TEST=Added a new unit test that stubs the validity check. Fuzzer failure
no longer reproduces after this patch.

Change-Id: I47af6bd65da277a652f4d0da40dd6ffa3edb1f0f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/2008272
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Tested-by: Dylan Reid <dgreid@chromium.org>
Commit-Queue: Dylan Reid <dgreid@chromium.org>