CRAS: dbus_control: check that dev_name is utf8
One of the most common causes of crashes for cras is due to a failed
internal verification step in libdbus when appending the device name to
a dbus message. In total there are over 120,000 such reports.

While the crash reporting is missing line number information within
dbus, we can guess that the cause is due to a non-utf8 string. In order
to hopefully reduce the odds of this crash, check that dev_name is a
utf8 string before appending it to the message. If it's not utf8, set it
to "".

BUG=None
TEST=unittests

Change-Id: If330d5ae875fbfebf55fa4bbb815cb67bb732ec6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/2038461
Commit-Queue: Fletcher Woodruff <fletcherw@chromium.org>
Tested-by: Fletcher Woodruff <fletcherw@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>